Extract `IAPIRequest` and `IAPIRequest<T>`
The initial reason for attempting this was that `APIRequest<T>` is
impossible to perform polymorphism over, as abstract classes do not
support generic type argument variance. Meanwhile, to use base interface
types as request responses, covariance support is needed, This was the
direct cause of adding `IAPIRequest<out T>`.

That, however, was quickly followed by the discovery that the above
basically necessitates creating the non-generic `IAPIRequest`, too.

The ugliest parts of this diff are the ones that interact with
`Trigger{Success,Failure}()` - the test-only `internal` helpers. Due to
language design reasons, calling into those methods after lifting them
out to the interface requires going explicitly through the interface
type itself, as implicit implementation of non-public interface members
is not allowed.