Fix beatmap set header border not being "inner alignment"
As said on Figma. Inner alignment can be done by separating the content with containers. Not sure about the performance implications. It also removes the horrid hover layer workaround.

It was outer alignment before, but something about the transparency is making the corner radius aliased. The difficulty drawables need it but is disabled right now.