Fix right clicks on timeline objects potentially getting eaten by playfield area
`SelectionHandler` is receiving input from anywhere out of necessity:

https://github.com/ppy/osu/blob/537e3111991c97018dc36c39faebe7ca683f5821/osu.Game/Screens/Edit/Compose/Components/SelectionHandler.cs#L119-L125

Also important is that `BlueprintContainer` will selectively not block
right clicks to make sure they fall through to the
`ContextMenuContainer`:

https://github.com/ppy/osu/blob/537e3111991c97018dc36c39faebe7ca683f5821/osu.Game/Screens/Edit/Compose/Components/BlueprintContainer.cs#L122-L126

But because the whole editor is sharing a `ContextMenuContainer` and
it's at a higher level than both components, we observe here the
playfield's `SelectionHandler` intercepting the right click before it
can reach the `ContextMenuContainer`.

The fix here is similar to what we're already doing in
`TimelineBlueprintContaienr`.