Fix collections dropdown crashing during storage migration
Closes https://github.com/ppy/osu/issues/25815.

`CollectionDropdown.collectionsChanged()` was assuming that if it
received `null` changes, then it must mean that the change subscription
is being initialised and the `filters` list will not contain any items.

However, that is not the only circumstance wherein a realm subscription
can fire with `null` changes; that can also happen after the main realm
instance gets recycled via the notification registration flow:

	https://github.com/ppy/osu/blob/9dffdaa0533ab8fe883df3f1d2e9ed9fe4e33f30/osu.Game/Database/RealmAccess.cs#L545-L549

	https://github.com/ppy/osu/blob/9dffdaa0533ab8fe883df3f1d2e9ed9fe4e33f30/osu.Game/Database/RealmAccess.cs#L1228-L1251

Therefore, to fix the crash, just ensure that the list is cleared every
time.