Fix osu! standardised score conversion sometimes exceeding bounds
Co-authored-by: Zyf <zyfarok@gmail.com>

Closes https://github.com/ppy/osu/issues/25860

Users reported that some stable scores would convert to large negative
total scores in lazer after the introduction of combo exponent. Those
large negative total scores were actually mangled NaNs.

The root cause of this was the following calculation going below zero
unexpectedly:

	https://github.com/ppy/osu/blob/739cafc86291799a36e61ebefae16b5c61d1de5d/osu.Game/Database/StandardisedScoreMigrationTools.cs#L323

which then propagates negative numbers onward until

	https://github.com/ppy/osu/blob/739cafc86291799a36e61ebefae16b5c61d1de5d/osu.Game/Database/StandardisedScoreMigrationTools.cs#L337

which yields a NaN due to attempting to take the square root of a
negative number.

To fix, clamp `comboPortionInScoreV1` to sane limits: to
`comboPortionFromLongestComboInScoreV1` from below, and to
`maximumAchievableComboPortionInScoreV1` from above. This is a less
direct fix than perhaps imagined, but it seems like a better one as it
will also affect the calculation of both the lower and the upper
estimate of the score.